pdff_coinverto_project/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdff_coinverto"
version = "0.1.0"
description = "A library to convert almost any file to PDF"
authors = [
  { name = "Antigravity", email = "antigravity@example.com" },
]
dependencies = [
  "Pillow",
  "fpdf",
  "docx2pdf; sys_platform == 'win32'",
  "openpyxl",
  "fastmcp",
  "markdown",
]
readme = "README.md"
requires-python = ">=3.8"

pdfitdown = "pdff_coinverto.server:main"
